# perf: improve performance through inline variable 'length'