--- a/lib/extend/filter.js
+++ b/lib/extend/filter.js
@@ -72,7 +72,7 @@ class Filter {
 
     args.unshift(data);
 
-    for (let i = 0, len = filters.length; i < len; i++) {
+    for (let i = 0; i < filters.length; i++) {
       const result = Reflect.apply(filters[i], ctx, args);
       args[0] = result == null ? args[0] : result;
     }