--- a/lib/hexo/index.js
+++ b/lib/hexo/index.js
@@ -43,14 +43,13 @@ const createLoadThemeRoute = function(generatorResult, locals, ctx) {
   const { path, cache: useCache } = locals;
 
   const layout = [...new Set(castArray(generatorResult.layout))];
-  const layoutLength = layout.length;
 
   // allways use cache in fragment_cache
   locals.cache = true;
   return () => {
     if (useCache && routeCache.has(generatorResult)) return routeCache.get(generatorResult);
 
-    for (let i = 0; i < layoutLength; i++) {
+    for (let i = 0; i < layout.length; i++) {
       const name = layout[i];
       const view = theme.getView(name);
 
@@ -424,7 +423,7 @@ class Hexo {
       // Remove old routes
       const removed = routeList.filter(item => !newRouteList.includes(item));
 
-      for (let i = 0, len = removed.length; i < len; i++) {
+      for (let i = 0; i < removed.length; i++) {
         route.remove(removed[i]);
       }
     });