--- a/lib/hexo/locals.js
+++ b/lib/hexo/locals.js
@@ -50,7 +50,7 @@ class Locals {
     const result = {};
     const keys = Object.keys(this.getters);
 
-    for (let i = 0, len = keys.length; i < len; i++) {
+    for (let i = 0; i < keys.length; i++) {
       const key = keys[i];
       const item = this.get(key);