--- a/lib/hexo/multi_config_path.js
+++ b/lib/hexo/multi_config_path.js
@@ -29,13 +29,10 @@ module.exports = ctx => function multiConfigPath(base, configPaths, outputDir) {
 
     return configPath;
   }
-
-  const numPaths = paths.length;
-
   // combine files
   let combinedConfig = {};
   let count = 0;
-  for (let i = 0; i < numPaths; i++) {
+  for (let i = 0; i < paths.length; i++) {
     const configPath = isAbsolute(paths[i]) ? paths[i] : join(base, paths[i]);
 
     if (!fs.existsSync(configPath)) {