--- a/lib/hexo/register_models.js
+++ b/lib/hexo/register_models.js
@@ -7,7 +7,7 @@ module.exports = ctx => {
 
   const keys = Object.keys(models);
 
-  for (let i = 0, len = keys.length; i < len; i++) {
+  for (let i = 0; i < keys.length; i++) {
     const key = keys[i];
     db.model(key, models[key](ctx));
   }