--- a/lib/plugins/console/config.js
+++ b/lib/plugins/console/config.js
@@ -41,7 +41,7 @@ function getProperty(obj, key) {
   const split = key.split('.');
   let result = obj[split[0]];
 
-  for (let i = 1, len = split.length; i < len; i++) {
+  for (let i = 1; i < split.length; i++) {
     result = result[split[i]];
   }
 
@@ -53,7 +53,7 @@ function setProperty(obj, key, value) {
   let cursor = obj;
   const lastKey = split.pop();
 
-  for (let i = 0, len = split.length; i < len; i++) {
+  for (let i = 0; i < split.length; i++) {
     const name = split[i];
     cursor[name] = cursor[name] || {};
     cursor = cursor[name];