--- a/lib/plugins/console/list/common.js
+++ b/lib/plugins/console/list/common.js
@@ -5,11 +5,10 @@ const strip = require('strip-ansi');
 exports.stringLength = str => {
   str = strip(str);
 
-  const len = str.length;
-  let result = len;
+  let result = str.length;
 
   // Detect double-byte characters
-  for (let i = 0; i < len; i++) {
+  for (let i = 0; i < str.length; i++) {
     if (str.charCodeAt(i) > 255) {
       result++;
     }