--- a/lib/plugins/console/list/route.js
+++ b/lib/plugins/console/list/route.js
@@ -19,11 +19,11 @@ function buildTree(routes) {
   const obj = {};
   let cursor;
 
-  for (let i = 0, len = routes.length; i < len; i++) {
+  for (let i = 0; i < routes.length; i++) {
     const item = routes[i].split('/');
     cursor = obj;
 
-    for (let j = 0, lenj = item.length; j < lenj; j++) {
+    for (let j = 0; j < item.length; j++) {
       const seg = item[j];
       cursor[seg] = cursor[seg] || {};
       cursor = cursor[seg];
@@ -37,7 +37,7 @@ function buildNodes(tree) {
   const keys = Object.keys(tree);
   const nodes = [];
 
-  for (let i = 0, len = keys.length; i < len; i++) {
+  for (let i = 0; i < keys.length; i++) {
     const key = keys[i];
     const item = tree[key];