--- a/lib/plugins/console/new.js
+++ b/lib/plugins/console/new.js
@@ -35,7 +35,7 @@ function newConsole(args) {
 
   const keys = Object.keys(args);
 
-  for (let i = 0, len = keys.length; i < len; i++) {
+  for (let i = 0; i < keys.length; i++) {
     const key = keys[i];
     if (!reservedKeys[key]) data[key] = args[key];
   }