--- a/lib/plugins/filter/new_post_path.js
+++ b/lib/plugins/filter/new_post_path.js
@@ -71,7 +71,7 @@ function newPostPathFilter(data = {}, replace) {
           hash
         };
 
-        for (let i = 0, len = keys.length; i < len; i++) {
+        for (let i = 0; i < keys.length; i++) {
           const key = keys[i];
           if (!reservedKeys[key]) filenameData[key] = data[key];
         }