--- a/lib/plugins/helper/list_archives.js
+++ b/lib/plugins/helper/list_archives.js
@@ -67,7 +67,7 @@ function listArchivesHelper(options = {}) {
   if (style === 'list') {
     result += `<ul class="${className}-list">`;
 
-    for (let i = 0, len = data.length; i < len; i++) {
+    for (let i = 0; i < data.length; i++) {
       const item = data[i];
 
       result += `<li class="${className}-list-item">`;
@@ -85,7 +85,7 @@ function listArchivesHelper(options = {}) {
 
     result += '</ul>';
   } else {
-    for (let i = 0, len = data.length; i < len; i++) {
+    for (let i = 0; i < data.length; i++) {
       const item = data[i];
 
       if (i) result += separator;