--- a/lib/plugins/helper/number_format.js
+++ b/lib/plugins/helper/number_format.js
@@ -15,7 +15,7 @@ function numberFormatHelper(num, options = {}) {
 
     if (beforeFirst) beforeArr.push(before.substr(0, beforeFirst));
 
-    for (let i = beforeFirst; i < beforeLength; i += 3) {
+    for (let i = beforeFirst; i < before.length; i += 3) {
       beforeArr.push(before.substr(i, 3));
     }
 
@@ -33,7 +33,7 @@ function numberFormatHelper(num, options = {}) {
       afterResult = after.substr(0, precision - 1) + (afterLast < 5 ? last : last + 1);
     } else {
       afterResult = after;
-      for (let i = 0, len = precision - afterLength; i < len; i++) {
+      for (let i = 0; i < precision - afterLength; i++) {
         afterResult += '0';
       }
     }