--- a/lib/plugins/helper/toc.js
+++ b/lib/plugins/helper/toc.js
@@ -22,7 +22,7 @@ function tocHelper(str, options = {}) {
   let firstLevel = 0;
   let lastLevel = 0;
 
-  for (let i = 0, len = data.length; i < len; i++) {
+  for (let i = 0; i < data.length; i++) {
     const el = data[i];
     const { level, id, text } = el;
     const href = id ? `#${id}` : id;