--- a/lib/plugins/processor/post.js
+++ b/lib/plugins/processor/post.js
@@ -60,7 +60,7 @@ module.exports = ctx => {
         data.published = false;
       }
 
-      for (let i = 0, len = keys.length; i < len; i++) {
+      for (let i = 0; i < keys.length; i++) {
         const key = keys[i];
         if (!preservedKeys[key]) data[key] = info[key];
       }