--- a/lib/plugins/tag/asset_img.js
+++ b/lib/plugins/tag/asset_img.js
@@ -14,10 +14,8 @@ module.exports = ctx => {
   const PostAsset = ctx.model('PostAsset');
 
   return function assetImgTag(args) {
-    const len = args.length;
-
     // Find image URL
-    for (let i = 0; i < len; i++) {
+    for (let i = 0; i < args.length; i++) {
       const asset = PostAsset.findOne({post: this._id, slug: args[i]});
       if (asset) {
         args[i] = encodeURL(resolve('/', asset.path));