```js
const obj = {}
;[...Array.from({ length: 10000 }).keys()].forEach(v => obj[v] = v)
const keys = Object.keys(obj)

console.time('1')
for (let i = 0, len = keys.length; i < len; i++) {
  const _ = 1
}
console.timeEnd('1')

console.time('2')
for (let i = 0; i < keys.length; i++) {
  const _ = 1
}
console.timeEnd('2')
```

```bash
C:\Users\Himself65\Desktop\github\test>node main.js
1: 0.242ms
2: 0.143ms

C:\Users\Himself65\Desktop\github\test>node main.js
1: 0.241ms
2: 0.153ms

C:\Users\Himself65\Desktop\github\test>node -v
v10.19.0
```

## Pull request tasks

- [X] Passed the CI test.